Fetch a Quantarctica data set

Fetch a Quantarctica data set

qa_get(dataset, cache_directory = qa_cache_dir(), refresh_cache = 0,
  verbose = FALSE, shapefile_reader, raster_reader)

Arguments

dataset
string or tibble: the name of the data set or a tibble as returned by qa_dataset

cache_directory
string: the cache directory to use. As for the path parameter to the qa_cache_dir function

refresh_cache
numeric: 0 = do not overwrite existing files, 1 = overwrite if the remote file is newer than the local copy, 2 = always overwrite existing files

verbose
logical: show progress messages?

shapefile_reader
function: function to use to read shapefiles. By default this is raster::shapefile

raster_reader
function: function to use to read raster files (TIFF, JP2, VRT). By default this is raster::raster

Value

By default, an object of class SpatialPolygonsDataFrame (for shapefile layers) or RasterLayer (for raster layers). Objects of other classes may be returned if non-default shapefile_reader or raster_reader functions are used