Has a dataset been downloaded to the local cache?

Has a dataset been downloaded to the local cache?

qa_is_cached(dataset, cache_directory = qa_cache_dir(), verbose = FALSE)

Arguments

dataset

string or qa_dataset: the dataset object, or name of the dataset

cache_directory

string: the cache directory to check. As for the path parameter to the qa_cache_dir function

verbose

logical: show progress messages?

Value

Logical

See also

Examples

if (FALSE) {
  ds <- qa_dataset("ADD Simple basemap")
  qa_is_cached(ds) ## FALSE
  x <- qa_get(ds)
  qa_is_cached(ds) ## TRUE
}